Termine el laboratorio en clase

+++ src/main/java/umg/programacion2/Formularios/frmProducto.java
@@ -2,10 +2,12 @@ package umg.programacion2.Formularios;
 
 import umg.programacion2.DataBase.Model.ProductoModel;
 import umg.programacion2.DataBase.Service.ProductoService;
+import umg.programacion2.Reportes.PdfReport;
 
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.List;
 
 public class frmProducto extends JFrame {
     private JPanel panel1;
@@ -21,6 +23,12 @@ public class frmProducto extends JFrame {
     private JButton buttonActualizar;
     private JButton buttonBorrar;
     private JButton buttonPDF;
+    private JLabel lblPrecio;
+    private JLabel lblExistencia;
+    private JTextField textFieldExistencia;
+    private JTextField textFieldPrecio;
+    private JLabel lblReportes;
+    private JComboBox comboBoxReportes;
 
     private ProductoService productoService;
 
@@ -31,7 +39,7 @@ public class frmProducto extends JFrame {
         // Configuración del JFrame
         setContentPane(panel1);
         setTitle("Gestión de Productos");
-        setSize(500, 300);
+        setSize(600, 300);
         setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
         setLocationRelativeTo(null);
 
@@ -53,6 +61,23 @@ public class frmProducto extends JFrame {
         comboBoxOrigen.addItem("China");
         comboBoxOrigen.addItem("Alemania");
         comboBoxOrigen.addItem("Escocia");
+
+        //Agregar al combo box reporet
+
+        // Definir las opciones en un array
+        String[] opcionesReportes = {
+                "Reporte General",
+                "Precio menores a 100",
+                "Existencia menor a 30 unidades",
+                "Precio entre 200 y 400",
+                "Ordenar de precio mayor a menor",
+                "Ordenar de  precio menor a mayor"
+        };
+
+        // Agregar las opciones al comboBox en un bucle
+        for (String opcion : opcionesReportes) {
+            comboBoxReportes.addItem(opcion);
+        }
 
         // Acción del botón Guardar
         buttonGuardar.addActionListener(new ActionListener() {
@@ -87,6 +112,9 @@ public class frmProducto extends JFrame {
                     if (producto != null) {
                         textFieldNombre.setText(producto.getDescripcion());
                         comboBoxOrigen.setSelectedItem(producto.getOrigen());
+                        textFieldExistencia.setText(String.valueOf(producto.getExistencia()));
+                        textFieldPrecio.setText(String.valueOf(producto.getPrecio()));
+
                     } else {
                         JOptionPane.showMessageDialog(null, "Producto no encontrado");
                     }
@@ -130,17 +158,69 @@ public class frmProducto extends JFrame {
                 // Borrar producto
                 try {
                     int idProducto = Integer.parseInt(textFieldCodigo.getText());
+                    ProductoModel producto = productoService.obtenerProductoPorId(idProducto);
                     int confirm = JOptionPane.showConfirmDialog(null, "¿Está seguro de eliminar este producto?", "Confirmación", JOptionPane.YES_NO_OPTION);
 
                     if (confirm == JOptionPane.YES_OPTION) {
-                        productoService.eliminarProducto(idProducto);
-                        JOptionPane.showMessageDialog(null, "Producto eliminado exitosamente");
-                        limpiarCampos();
+                        if(producto.getPrecio() == 0) {
+                            productoService.eliminarProducto(idProducto);
+                            JOptionPane.showMessageDialog(null, "Producto eliminado exitosamente");
+                            limpiarCampos();
+                        }else
+                        {
+                            JOptionPane.showMessageDialog(null, "Error no se puede eliminar un prodcuto si su precio no es de Q0.00");
+                        }
                     }
                 } catch (NumberFormatException ex) {
                     JOptionPane.showMessageDialog(null, "Código inválido");
                 } catch (Exception ex) {
                     JOptionPane.showMessageDialog(null, "Error al eliminar el producto: " + ex.getMessage());
+                }
+            }
+        });
+        buttonPDF.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                try{
+
+                    int ReporteSelccionado = comboBoxReportes.getSelectedIndex();
+                    switch (ReporteSelccionado) {
+                        case 0:
+                            // Acción para "Reporte General"
+                            List<ProductoModel> prod = new ProductoService().obtenerTodosLosProductos();
+                            new PdfReport().generateProductReport(prod,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                        case 1:
+                            // Acción para "Precio menores a 100"
+                            List<ProductoModel> prod2 = new ProductoService().obtenerGenericos("precio < 100");
+                            new PdfReport().generateProductReport(prod2,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                        case 2:
+                            // Acción para "Existencia menor a 30 unidades"
+                            List<ProductoModel> prod3 = new ProductoService().obtenerGenericos("existencia < 30");
+                            new PdfReport().generateProductReport(prod3,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                        case 3:
+                            // Acción para "Precio entre 200 y 400"
+                            List<ProductoModel> prod4 = new ProductoService().obtenerGenericos("precio BETWEEN 200 AND 400");
+                            new PdfReport().generateProductReport(prod4,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                        case 4:
+                            // Acción para "Ordenar de mayor a menor"
+                            List<ProductoModel> prod5 = new ProductoService().obtenerGenericos("1=1 ORDER BY precio DESC");
+                            new PdfReport().generateProductReport(prod5,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                        case 5:
+                            // Acción para "Ordenar de menor a mayor"
+                            List<ProductoModel> prod6 = new ProductoService().obtenerGenericos("1=1 ORDER BY precio ASC");
+                            new PdfReport().generateProductReport(prod6,"C:\\PdfProgra\\reporte.pdf");
+                            break;
+                    }
+
+                    JOptionPane.showMessageDialog(null, "Reporte generado en C:\\PdfProgra");
+                }catch (Exception exception)
+                {
+                    System.out.println("Error: " + exception.getMessage());
                 }
             }
         });
